+++ lib/utils/copy_clipboard.dart
@@ -3,8 +3,12 @@ import 'package:flutter/services.dart';
 
 ///复制到粘贴板
 copyClipboard(BuildContext context, String? value) {
-  var snackBar =
-      SnackBar(content: Text('$value\n\n copy success to clipboard'));
-  ScaffoldMessenger.of(context).showSnackBar(snackBar);
+  snackBar(context,'$value\n\n copy success to clipboard');
   Clipboard.setData(ClipboardData(text: value ?? 'null'));
+}
+
+snackBar(BuildContext context, String? value){
+  var snackBar =
+      SnackBar(content: Text('$value'));
+  ScaffoldMessenger.of(context).showSnackBar(snackBar);
 }